requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
deepgram-sdk==3.7.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
elevenlabs==1.13.4
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
moviepy==2.1.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.0
openai==1.57.1
packaging==24.2
pillow==10.4.0
postgrest==0.18.0
proglog==0.1.10
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
realtime==2.0.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==13.1
yarl==1.18.3
